Ask60 to master

See merge request compsci308_2020spring/parser_parser_team14!17

+++ src/slogo/model/backEndInternal/commands/PenDown.java
@@ -1,9 +1,8 @@
 package slogo.model.backEndInternal.commands;
 
 import slogo.model.backEndInternal.BackEndTurtle;
-import slogo.model.backEndInternal.commands.Command;
 
-public class PenDown implements Command<Integer> {
+public class PenDown implements Command<Double> {
 
   BackEndTurtle backEndTurtle;
 
@@ -12,8 +11,8 @@ public class PenDown implements Command<Integer> {
   }
 
   @Override
-  public Integer execute() {
+  public Double execute() {
     backEndTurtle.flipPen();
-    return 1;
+    return 1.0;
   }
 }
